Adding a throw away change to models
The tutorial talks about adding a throw away change to test downgrades.  It doesn't work for me, can't make the migration work because it tells me the database is not up to date. Without migrations I don't know how it is possible for the database to be up to date.  Apparently the tutorial is working from an old version of flask-migrate, need to learn about the current version and how it works (with revision, migrate, upgrade and downgrade commands).